renamed metplus/util/string to string_manip because the name was clashing with the std string library when running the diff tests, ci-run-all-diff
