+++ metplus/util/string_manip.py
